requirements.txt
python-telegram-bot==21.9
openai>=1.40.0
pypdf==5.1.0
python-docx==1.1.2
python-dotenv==1.2.2
black==26.1.0